RMSRCBUILD-13 Supply a local nuspec-file for projects that use a referenced nuspec-file to support dependent projects.
The NuGet command line tool uses a very specific set of rules when creating a package for a projects that has local dependencies. In short, it requires a local nuspec file named like the referenced project. This nuspec-file only supplies the version number and the identity of the referenced project. Due to some inconsistency in the NuGet command line tool, the nuspec-file referenced by the project cannot be used for this.

In addition to this peculiarity, the nuspec-file referenced by the project must not define any hard file-dependencies. Otherwise, the NuGet command line tool will try to look up the file-reference based on the current project's base path and failing because it cannot be found.