RMLNQSQL-82 Change support for tags to use AssemblyMetadataAttribute.
In NuGet 3.0, the AssemblyMetadataAttribute will be used to communicate arbitrary properties to the NuGet pack command. Until the feature becomes generally available, the build script will extract the tags-metadata from AssemblyInfo.cs files located in the Properties folder. This implementation covers the primary usecase of the build script to create NuGet packages from C# projects. This feature is only made available for Release-projects, not for Test-projects.